requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
pandas==2.2.3
openpyxl==3.1.5
python-multipart==0.0.20
jinja2==3.1.5
pyyaml==6.0.2
tqdm==4.67.1
